+++ src/equations/implicit_equation.h
@@ -9,7 +9,7 @@ public:
     ImplicitEquation(DisplaySettings settings, float (*func)(float,float));
     float apply(float x, float y); // Function drawn where apply(x,y)=0
 
-    GLfloat* getVertices(int& numVerts, BoundingBox boundingBox, float precision) override;
+    GLfloat* getVertices(unsigned long& numVerts, BoundingBox boundingBox, float precision) override;
 
 private:
     float (*function)(float,float);
